Foundation of the Google Drive project for Eclipse.
The goal is to allow users to backup and sync selected projects using
Google Drive.

This allow users to start working in new instances of Eclipse without
having to manually resetup projects every time they install a new
instance of Eclipse. 

Syncing allows the same project in all instances of Eclipse to be in 
sync.